+++ pkg/collector/process_test.go
@@ -44,7 +44,7 @@ func TestTCPCollectingProcess_ReceiveTemplateRecord(t *testing.T) {
 	if err != nil {
 		t.Error(err)
 	}
-	cp, err := InitCollectingProcess(address, 1024, 0, nil)
+	cp, err := InitCollectingProcess(address, 1024, 0, false)
 	if err != nil {
 		t.Fatalf("TCP Collecting Process does not start correctly: %v", err)
 	}
@@ -70,7 +70,7 @@ func TestUDPCollectingProcess_ReceiveTemplateRecord(t *testing.T) {
 	if err != nil {
 		t.Error(err)
 	}
-	cp, err := InitCollectingProcess(address, 1024, 0, nil)
+	cp, err := InitCollectingProcess(address, 1024, 0, false)
 	if err != nil {
 		t.Fatalf("UDP Collecting Process does not start correctly: %v", err)
 	}
@@ -100,9 +100,8 @@ func TestTCPCollectingProcess_ReceiveDataRecord(t *testing.T) {
 	if err != nil {
 		t.Error(err)
 	}
-	messageChan := make(chan *entities.Message)
 	messageCount := 0
-	cp, err := InitCollectingProcess(address, 1024, 0, messageChan)
+	cp, err := InitCollectingProcess(address, 1024, 0, true)
 	// Add the templates before sending data record
 	cp.addTemplate(uint32(1), uint16(256), elementsWithValue)
 	if err != nil {
@@ -116,7 +115,7 @@ func TestTCPCollectingProcess_ReceiveDataRecord(t *testing.T) {
 		}
 		defer conn.Close()
 		conn.Write(validDataPacket)
-		for range messageChan {
+		for range cp.GetMsgChan() {
 			messageCount++
 		}
 	}()
@@ -133,9 +132,8 @@ func TestUDPCollectingProcess_ReceiveDataRecord(t *testing.T) {
 	if err != nil {
 		t.Error(err)
 	}
-	messageChan := make(chan *entities.Message)
 	messageCount := 0
-	cp, err := InitCollectingProcess(address, 1024, 0, messageChan)
+	cp, err := InitCollectingProcess(address, 1024, 0, true)
 	// Add the templates before sending data record
 	cp.addTemplate(uint32(1), uint16(256), elementsWithValue)
 	if err != nil {
@@ -153,7 +151,7 @@ func TestUDPCollectingProcess_ReceiveDataRecord(t *testing.T) {
 		}
 		defer conn.Close()
 		conn.Write(validDataPacket)
-		for range messageChan {
+		for range cp.GetMsgChan() {
 			messageCount++
 		}
 	}()
@@ -170,7 +168,7 @@ func TestTCPCollectingProcess_ConcurrentClient(t *testing.T) {
 	if err != nil {
 		t.Error(err)
 	}
-	cp, _ := InitCollectingProcess(address, 1024, 0, nil)
+	cp, _ := InitCollectingProcess(address, 1024, 0, false)
 	go func() {
 		time.Sleep(time.Second)
 		_, err := net.Dial(address.Network(), address.String())
@@ -196,7 +194,7 @@ func TestUDPCollectingProcess_ConcurrentClient(t *testing.T) {
 	if err != nil {
 		t.Error(err)
 	}
-	cp, _ := InitCollectingProcess(address, 1024, 0, nil)
+	cp, _ := InitCollectingProcess(address, 1024, 0, false)
 	go func() {
 		time.Sleep(time.Second)
 		resolveAddr, err := net.ResolveUDPAddr(address.Network(), address.String())
@@ -295,8 +293,6 @@ func TestCollectingProcess_DecodeDataRecord(t *testing.T) {
 	ipAddress := net.IP([]byte{1, 2, 3, 4})
 	elements := v.GetRecords()[0].GetInfoElements()
 	assert.Equal(t, ipAddress, elements[0].Value, "sourceIPv4Address should be decoded and stored correctly.")
-	assert.Equal(t, uint32(0), elements[3].Value, "originalExporterIPv4Address should be added to record correctly.")
-	assert.Equal(t, uint32(1), elements[4].Value, "originalObservationDomainId should be added to record correctly.")
 	// Malformed data record
 	dataRecord := []byte{0, 10, 0, 33, 95, 40, 212, 159, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0}
 	_, err = cp.decodePacket(bytes.NewBuffer(dataRecord), address.String())
@@ -308,7 +304,7 @@ func TestUDPCollectingProcess_TemplateExpire(t *testing.T) {
 	if err != nil {
 		t.Error(err)
 	}
-	cp, err := InitCollectingProcess(address, 1024, 5, nil)
+	cp, err := InitCollectingProcess(address, 1024, 5, false)
 	if err != nil {
 		t.Fatalf("UDP Collecting Process does not start correctly: %v", err)
 	}
@@ -336,63 +332,4 @@ func TestUDPCollectingProcess_TemplateExpire(t *testing.T) {
 	assert.NotNil(t, cp.templatesMap[1][256], "Template should be stored in the template map.")
 	time.Sleep(10 * time.Second)
 	assert.Nil(t, cp.templatesMap[1][256], "Template should be deleted after 5 seconds.")
-}
-
-func TestAddOriginalExporterInfo(t *testing.T) {
-	// Test message with template set
-	message := createMsgwithTemplateSet()
-	addOriginalExporterInfo(message)
-	record := message.Set.GetRecords()[0]
-	assert.Equal(t, "originalExporterIPv4Address", record.GetInfoElements()[5].Element.Name)
-	assert.Equal(t, "originalObservationDomainId", record.GetInfoElements()[6].Element.Name)
-	// Test message with data set
-	message = createMsgwithDataSet()
-	addOriginalExporterInfo(message)
-	record = message.Set.GetRecords()[0]
-	assert.Equal(t, "originalExporterIPv4Address", record.GetInfoElements()[5].Element.Name)
-	assert.Equal(t, uint32(2130706433), record.GetInfoElements()[5].Value)
-	assert.Equal(t, "originalObservationDomainId", record.GetInfoElements()[6].Element.Name)
-	assert.Equal(t, uint32(1234), record.GetInfoElements()[6].Value)
-}
-
-func createMsgwithTemplateSet() *entities.Message {
-	set := entities.NewSet(entities.Template, 256, false)
-	elements := make([]*entities.InfoElementWithValue, 0)
-	ie1 := entities.NewInfoElementWithValue(entities.NewInfoElement("sourceIPv4Address", 8, 18, 0, 4), nil)
-	ie2 := entities.NewInfoElementWithValue(entities.NewInfoElement("destinationIPv4Address", 12, 18, 0, 4), nil)
-	ie3 := entities.NewInfoElementWithValue(entities.NewInfoElement("sourceTransportPort", 7, 2, 0, 2), nil)
-	ie4 := entities.NewInfoElementWithValue(entities.NewInfoElement("destinationTransportPort", 11, 2, 0, 2), nil)
-	ie5 := entities.NewInfoElementWithValue(entities.NewInfoElement("protocolIdentifier", 4, 1, 0, 1), nil)
-	elements = append(elements, ie1, ie2, ie3, ie4, ie5)
-	set.AddRecord(elements, 256)
-	return &entities.Message{
-		Version:       10,
-		BufferLength:  40,
-		SeqNumber:     1,
-		ObsDomainID:   5678,
-		ExportTime:    0,
-		ExportAddress: "127.0.0.1",
-		Set:           set,
-	}
-}
-
-func createMsgwithDataSet() *entities.Message {
-	set := entities.NewSet(entities.Data, 256, false)
-	elements := make([]*entities.InfoElementWithValue, 0)
-	ie1 := entities.NewInfoElementWithValue(entities.NewInfoElement("sourceIPv4Address", 8, 18, 0, 4), net.ParseIP("10.0.0.1"))
-	ie2 := entities.NewInfoElementWithValue(entities.NewInfoElement("destinationIPv4Address", 12, 18, 0, 4), net.ParseIP("10.0.0.2"))
-	ie3 := entities.NewInfoElementWithValue(entities.NewInfoElement("sourceTransportPort", 7, 2, 0, 2), uint16(1234))
-	ie4 := entities.NewInfoElementWithValue(entities.NewInfoElement("destinationTransportPort", 11, 2, 0, 2), uint16(5678))
-	ie5 := entities.NewInfoElementWithValue(entities.NewInfoElement("protocolIdentifier", 4, 1, 0, 1), uint8(6))
-	elements = append(elements, ie1, ie2, ie3, ie4, ie5)
-	set.AddRecord(elements, 256)
-	return &entities.Message{
-		Version:       10,
-		BufferLength:  32,
-		SeqNumber:     1,
-		ObsDomainID:   uint32(1234),
-		ExportTime:    0,
-		ExportAddress: "127.0.0.1",
-		Set:           set,
-	}
 }
